# backend/docs/swagger.yaml
definitions:
...
  entity.Key:
    properties:
      id:
        example: 1
        type: integer
      publicKey:
        example: GCK...
        type: string
      walletId:
        example: 1
        type: integer
      weight:
        example: 3
        type: integer
    type: object
...
  entity.Vault:
    properties:
      active:
        type: integer
      id:
        example: 1
        type: integer
      name:
        example: Treasury
        type: string
      vault_category:
        $ref: '#/definitions/entity.VaultCategory'
      wallet:
        $ref: '#/definitions/entity.Wallet'
    type: object
  entity.VaultCategory:
    properties:
      id:
        example: 1
        type: integer
      name:
        example: Treasury
        type: string
      theme:
        example: blue
        type: string
    type: object
  entity.Wallet:
    properties:
      funded:
        type: boolean
      id:
        example: 1
        type: integer
      key:
        $ref: '#/definitions/entity.Key'
      type:
        example: sponsor
        type: string
    type: object
...
  v1.CreateVaultRequest:
    properties:
      assets_id:
        items:
          type: integer
        type: array
      name:
        example: Treasury
        type: string
      vault_category_id:
        example: 1
        type: integer
    required:
    - assets_id
    - name
    - vault_category_id
    type: object
...
  v1.UpdateVaultAssetRequest:
    properties:
      asset_code:
        example: "1"
        type: string
      asset_issuer_pk:
        example: "1"
        type: string
      is_add:
        example: true
        type: boolean
      is_remove:
        example: false
        type: boolean
    required:
    - asset_code
    - asset_issuer_pk
    type: object
  v1.UpdateVaultCategoryRequest:
    properties:
      name:
        example: Vault 2
        type: string
      vault_category_id:
        example: 1
        type: integer
    required:
    - vault_category_id
    type: object
  v1.response:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /vault:
    post:
      consumes:
      - application/json
      description: Create and issue a new asset on Stellar
      parameters:
      - description: CreateVaultRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.CreateVaultRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Vault'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Create a new vault
      tags:
      - Vault
  /vault-asset/{id}:
    put:
      consumes:
      - application/json
      description: Update a vault by providing the Vault ID and the updated the asset.
      parameters:
      - description: Vault ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Vault asset info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateVaultAssetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated vault asset information
          schema:
            $ref: '#/definitions/entity.Vault'
        "400":
          description: 'Bad Request: Invalid input data'
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: 'Not Found: Vault not found'
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: 'Internal Server Error: Failed to update vault asset'
          schema:
            $ref: '#/definitions/v1.response'
      summary: Update a vault asset
      tags:
      - Vault
...
  /vault-category/{id}:
    put:
      consumes:
      - application/json
      description: Update a vault category by providing the Vault Category ID and
        the updated information.
      parameters:
      - description: Vault Category ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Vault Category info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateVaultCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated vault category information
          schema:
            $ref: '#/definitions/entity.Vault'
        "400":
          description: 'Bad Request: Invalid input data'
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: 'Not Found: Vault category not found'
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: 'Internal Server Error: Failed to update vault category'
          schema:
            $ref: '#/definitions/v1.response'
      summary: Update a vault category
      tags:
      - Vault Category
  /vault-delete/{id}:
    put:
      consumes:
      - application/json
      description: Update a vault by providing the Vault ID and the updated the status.
      parameters:
      - description: Vault ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Updated vault status information
          schema:
            $ref: '#/definitions/entity.Vault'
        "400":
          description: 'Bad Request: Invalid input data'
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: 'Not Found: Vault not found'
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: 'Internal Server Error: Failed to update vault status'
          schema:
            $ref: '#/definitions/v1.response'
      summary: Update a vault status
      tags:
      - Vault
  /vault/{id}:
    get:
      consumes:
      - application/json
      description: Get vault
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Vault'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Get vault
      tags:
      - Vault
  /vault/list:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all vaults, with optional pagination
      parameters:
      - description: Page number for pagination
        in: query
        name: page
        type: integer
      - description: Number of items per page for pagination
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Vault'
            type: array
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Get all vaults
      tags:
      - Vault
...
swagger: "2.0"